requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
PyYAML
Flask
APScheduler
SQLAlchemy
python-dotenv